ref(ui-components) Replace Switch with new components (#11956)
